Add ActiveEnrollmentSet, study_admin group/account
ActiveEnrollmentSet is used to keep track of what current enrollment
processes should be used by the site - e.g. the requirements list it
should display.

To limit the number of ActiveEnrollmentSets to one, one is created by
the fixtures. The study_admin group is created that can edit it (but
cannot add or delet an ActiveEnrollmentSet) as well as edit the other
enrollment components within it.

One user is created called "study_admin" in the study_admin group,
with the default password as "study_admin". Setting up
django_study_enrollment should be streamlined: the user can immediately
log in to this account and use it to create and administer the desired
enrollment process.